add md/verify

+++ math/pow.cpp
@@ -1,6 +1,6 @@
 template <class T>
 T pow_ (T x, T n, T M){
-    uint64_t u = 1, xx = x;
+    unsigned __int128 u = 1, xx = x;
     while (n > 0){
         if (n&1) u = u * xx % M;
         xx = xx * xx % M;
